+++ src/main.jsx
@@ -5,7 +5,11 @@ import { HelmetProvider } from "react-helmet-async";
 import { RouterProvider } from "react-router-dom";
 import { ErrorFallback } from "./components/ui/animation";
 import { ExternalLinkProvider } from "./features/external-link/context/ExternalLinkContext";
-import { PerformanceProvider, ThemeProvider } from "./features/settings";
+import {
+  AccessibilityProvider,
+  PerformanceProvider,
+  ThemeProvider,
+} from "./features/settings";
 import "./index.css";
 import router from "./routes/Router.jsx";
 
@@ -15,9 +19,11 @@ createRoot(document.getElementById("root")).render(
       <ErrorBoundary FallbackComponent={ErrorFallback}>
         <ThemeProvider>
           <PerformanceProvider>
-            <ExternalLinkProvider>
-              <RouterProvider router={router} />
-            </ExternalLinkProvider>
+            <AccessibilityProvider>
+              <ExternalLinkProvider>
+                <RouterProvider router={router} />
+              </ExternalLinkProvider>
+            </AccessibilityProvider>
           </PerformanceProvider>
         </ThemeProvider>
       </ErrorBoundary>